# src/main/resources/META-INF/resources/openapi.yaml
...
openapi: 3.0.3
info:
  title: parcel-sorting-machine-service API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:9000
  description: Auto generated value
- url: http://0.0.0.0:9000
  description: Auto generated value
paths:
...
  /parcel/container/open/{id}:
    get:
      tags:
      - App Ressource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
...
  /parcel/item/{rec}:
    get:
      tags:
      - App Ressource
      parameters:
      - name: rec
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
...